forecasting/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.4=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.1.0=py38hecd8cb5_0
backcall=0.2.0=pyh9f0ad1d_0
brotli=1.1.0=h0dc2134_1
brotli-bin=1.1.0=h0dc2134_1
ca-certificates=2024.3.11=hecd8cb5_0
cctools_osx-64=973.0.1=habff3f6_15
certifi=2024.2.2=py38hecd8cb5_0
clang=15.0.7=hdae98eb_5
clang-15=15.0.7=default_h7151d67_5
clang_impl_osx-64=15.0.7=h03d6864_8
clang_osx-64=15.0.7=hb91bd55_8
clangxx=15.0.7=default_h7151d67_5
clangxx_impl_osx-64=15.0.7=h2133e9c_8
clangxx_osx-64=15.0.7=hb91bd55_8
cloudpickle=3.0.0=pypi_0
cmdstan=2.33.1=ha749d2a_0
cmdstanpy=1.2.2=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
compiler-rt=15.0.7=ha38d28d_2
compiler-rt_osx-64=15.0.7=ha38d28d_2
contourpy=1.1.1=py38h15a1a5b_1
convertdate=2.4.0=pyhd8ed1ab_0
cycler=0.12.1=pyhd8ed1ab_0
debugpy=1.8.1=py38h1f5f77c_0
decorator=5.1.1=pyhd8ed1ab_0
ephem=4.1.5=py38hae2e43d_1
executing=2.0.1=pyhd8ed1ab_0
fonttools=4.51.0=py38hae2e43d_0
freetype=2.12.1=h60636b9_2
future=1.0.0=pypi_0
holidays=0.46=pyhd8ed1ab_0
hyperopt=0.2.7=pypi_0
icu=73.2=hf5e326d_0
importlib-metadata=7.1.0=pyha770c72_0
importlib-resources=6.4.0=pyhd8ed1ab_0
importlib_metadata=7.1.0=hd8ed1ab_0
importlib_resources=6.4.0=pyhd8ed1ab_0
ipykernel=6.29.3=pyh3cd1d5f_0
ipython=8.12.2=pyhd1c38e8_0
jedi=0.19.1=pyhd8ed1ab_0
joblib=1.4.0=pypi_0
jsonschema=4.19.2=py38hecd8cb5_0
jsonschema-specifications=2023.7.1=py38hecd8cb5_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_core=5.7.2=py38h50d1736_0
kiwisolver=1.4.5=py38h15a1a5b_1
lcms2=2.15=ha53face_2
ld64_osx-64=609=h0fd476b_15
lerc=4.0.0=hb486fe8_0
libblas=3.9.0=22_osx64_openblas
libbrotlicommon=1.1.0=h0dc2134_1
libbrotlidec=1.1.0=h0dc2134_1
libbrotlienc=1.1.0=h0dc2134_1
libcblas=3.9.0=22_osx64_openblas
libclang-cpp15=15.0.7=default_h7151d67_5
libcxx=16.0.6=hd57cbcb_0
libdeflate=1.19=ha4e1b8e_0
libffi=3.4.4=hecd8cb5_0
libgfortran=5.0.0=13_2_0_h97931a8_3
libgfortran5=13.2.0=h2873a65_3
libhwloc=2.9.3=default_h24e0189_1009
libiconv=1.17=hd75f5a5_2
libjpeg-turbo=2.1.5.1=h0dc2134_1
liblapack=3.9.0=22_osx64_openblas
libllvm15=15.0.7=hbedff68_4
libopenblas=0.3.27=openmp_hfef2a42_0
libpng=1.6.43=h92b6c6a_0
libsodium=1.0.18=hbcb3906_1
libtiff=4.6.0=haeeb97c_1
libwebp-base=1.3.2=h10d778d_1
libxcb=1.15=hb7f2c08_0
libxml2=2.12.6=hc0ae0f7_1
libzlib=1.2.13=h8a1eda9_5
llvm-openmp=18.1.3=hb6ac08f_0
llvm-tools=15.0.7=hbedff68_4
lunarcalendar=0.0.9=py_0
make=4.3=h22f3db7_1
matplotlib-base=3.7.3=py38hcd1b199_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
nbformat=5.9.2=py38hecd8cb5_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=pypi_0
numpy=1.24.4=py38h9a4a08f_0
openjpeg=2.5.2=h7310d3a_0
openssl=3.2.1=hd75f5a5_1
packaging=24.0=pyhd8ed1ab_0
pandas=2.0.3=py38h78e6021_1
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.0.1=py38ha0d9280_1
pip=23.3.1=py38hecd8cb5_0
pkgutil-resolve-name=1.3.10=py38hecd8cb5_1
platformdirs=4.2.0=pyhd8ed1ab_0
plotly=5.19.0=py38h20db666_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
prophet=1.1.5=py38h4405de9_1
psutil=5.9.8=py38hae2e43d_0
pthread-stubs=0.4=hc929b4f_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py4j=0.10.9.7=pypi_0
pyarrow=15.0.2=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pymeeus=0.5.12=pyhd8ed1ab_0
pyparsing=3.1.2=pyhd8ed1ab_0
pyspark=3.5.1=pypi_0
python=3.8.19=h5ee71fb_0
python-dateutil=2.9.0=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=py38hecd8cb5_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2024.1=pyhd8ed1ab_0
pyzmq=25.1.2=py38ha51bd64_0
readline=8.2=hca72f7f_0
referencing=0.30.2=py38hecd8cb5_0
regex=2023.12.25=pypi_0
rpds-py=0.10.6=py38hf2ad997_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=68.2.2=py38hecd8cb5_0
sigtool=0.1.3=h88f4db0_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h6c40b1e_0
stack_data=0.6.2=pyhd8ed1ab_0
stanio=0.5.0=pyhd8ed1ab_0
tapi=1100.0.11=h9ce4665_0
tbb=2021.11.0=h7728843_1
tbb-devel=2021.11.0=ha56d2d7_1
tenacity=8.2.2=py38hecd8cb5_0
threadpoolctl=3.4.0=pypi_0
tk=8.6.12=h5d9f67b_0
tornado=6.4=py38hae2e43d_0
tqdm=4.66.2=pyhd8ed1ab_0
traitlets=5.14.2=pyhd8ed1ab_0
typing_extensions=4.11.0=pyha770c72_0
unicodedata2=15.1.0=py38hcafd530_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.41.2=py38hecd8cb5_0
xorg-libxau=1.0.11=h0dc2134_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.4.6=h6c40b1e_0
zeromq=4.3.5=h93d8f39_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8a1eda9_5
zstd=1.5.5=h829000d_0
google-cloud-bigquery
